refactore: パッケージ名StudentManagementをstudent.managementに変更
