+++ src/main/java/raisetech/student/management/ServletInitializer.java
@@ -1,13 +1,13 @@
-package raisetech.StudentManagement;
+package raisetech.student.management;
 
 import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
 
 public class ServletInitializer extends SpringBootServletInitializer {
 
-	@Override
-	protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
-		return application.sources(StudentManagementApplication.class);
-	}
+  @Override
+  protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
+    return application.sources(StudentManagementApplication.class);
+  }
 
 }
