+++ src/main/java/raisetech/student/management/StudentManagementApplication.java
@@ -1,4 +1,4 @@
-package raisetech.StudentManagement;
+package raisetech.student.management;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
