+++ src/main/java/raisetech/student/management/controller/StudentController.java
@@ -1,4 +1,4 @@
-package raisetech.StudentManagement.controller;
+package raisetech.student.management.controller;
 
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Positive;
@@ -12,12 +12,12 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import raisetech.StudentManagement.data.domain.StudentDetail;
-import raisetech.StudentManagement.data.domain.validation.OnCreate;
-import raisetech.StudentManagement.data.domain.validation.OnUpdate;
-import raisetech.StudentManagement.exception.InvalidAccessException;
-import raisetech.StudentManagement.exception.NotExistIdException;
-import raisetech.StudentManagement.service.StudentService;
+import raisetech.student.management.data.domain.StudentDetail;
+import raisetech.student.management.data.domain.validation.OnCreate;
+import raisetech.student.management.data.domain.validation.OnUpdate;
+import raisetech.student.management.exception.InvalidAccessException;
+import raisetech.student.management.exception.NotExistIdException;
+import raisetech.student.management.service.StudentService;
 
 /**
  * 受講生の検索や登録、更新などを行うREST APIとして実行されるControllerです。
