+++ src/main/java/raisetech/student/management/data/domain/validation/OnCreate.java
@@ -1,4 +1,4 @@
-package raisetech.StudentManagement.data.domain.validation;
+package raisetech.student.management.data.domain.validation;
 
 /**
  * 新規登録時におけるバリデーション用グループ
