+++ src/main/java/raisetech/student/management/data/domain/validation/OnUpdate.java
@@ -1,4 +1,4 @@
-package raisetech.StudentManagement.data.domain.validation;
+package raisetech.student.management.data.domain.validation;
 
 /**
  * 更新時のバリデーション用グループ
