+++ src/main/java/raisetech/student/management/data/domain/validation/PhoneNumber.java
@@ -1,4 +1,4 @@
-package raisetech.StudentManagement.data.domain.validation;
+package raisetech.student.management.data.domain.validation;
 
 import jakarta.validation.Constraint;
 import jakarta.validation.Payload;
@@ -9,12 +9,12 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 @Documented
-@Constraint(validatedBy = CourseNameValidator.class)
+@Constraint(validatedBy = PhoneNumberValidator.class)
 @Target({ElementType.FIELD})
 @Retention(RetentionPolicy.RUNTIME)
-public @interface CourseName {
+public @interface PhoneNumber {
 
-  String message() default "コース名は「Javaコース」, 「AWSコース」,「デザインコース」,「Webマーケティングコース」,「フロントエンドコース」のいずれかを入力してください";
+  String message() default "電話番号の形式が不正です（例: 090-1234-5678）";
 
   Class<?>[] groups() default {};
