+++ src/main/java/raisetech/student/management/exception/InvalidAccessException.java
@@ -1,4 +1,4 @@
-package raisetech.StudentManagement.exception;
+package raisetech.student.management.exception;
 
 public class InvalidAccessException extends Exception {
