+++ src/main/java/raisetech/student/management/repository/StudentRepository.java
@@ -1,9 +1,9 @@
-package raisetech.StudentManagement.repository;
+package raisetech.student.management.repository;
 
 import java.util.List;
 import org.apache.ibatis.annotations.Mapper;
-import raisetech.StudentManagement.data.Student;
-import raisetech.StudentManagement.data.StudentCourse;
+import raisetech.student.management.data.Student;
+import raisetech.student.management.data.StudentCourse;
 
 /**
  * 受講生テーブルと受講生コーステーブルと紐づくRepositoryです。
