+++ src/main/java/raisetech/student/management/service/StudentService.java
@@ -1,4 +1,4 @@
-package raisetech.StudentManagement.service;
+package raisetech.student.management.service;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -6,10 +6,10 @@ import java.util.stream.Collectors;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import raisetech.StudentManagement.data.Student;
-import raisetech.StudentManagement.data.StudentCourse;
-import raisetech.StudentManagement.data.domain.StudentDetail;
-import raisetech.StudentManagement.repository.StudentRepository;
+import raisetech.student.management.data.Student;
+import raisetech.student.management.data.StudentCourse;
+import raisetech.student.management.data.domain.StudentDetail;
+import raisetech.student.management.repository.StudentRepository;
 
 /**
  * 受講生情報を取り扱うServiceです。受講生の検索や登録・更新処理を行います。
